+++ test_all.sh
@@ -1,0 +1,5 @@
+set -ex
+
+python MDF.py
+python Simple.py
+python ABCD.py
